minimal translation
After fixing the bugs, I created a minimal translation.

It is based on the Encyclopedia of Civilization.
Most of the names and descriptions of in-game mechanics have been translated.
Given my skills, the quality of the translation was not good, and there was a lot of transliteration and paraphrase, but since there was none, I just spent my time.
Additional elements such as work slots or trade routes have not been translated.

Also, sentences that overlap with vanilla, which also caused bugs. I checked in-game and deleted everything that was translated even if it wasn't there.

Lastly, purely explanatory sentences, city names, diplomatic lines, etc. were not translated.